Split compose package by purpose
